Enhance authentication
Previously, an instance of the project view model which has been
loaded from the database or the presentation model was used to check
the user's access privilege. This allowed an attacker to determine
if a certain ID is valid or not since different status codes were
returned.

With this commit a new empty project view model is created for
authorization checks which prevents such a leakage of information.

Furthermore, an authorization for the project list page has been added.